Limit the scope of profiler (#31)
In main.cpp, the destructor of the profiler was called after
amrex::Finalize.  This might cause an error in the future in SYCL if there
is a device synchronization call in the dtor, because the SYCL queues in
amrex have been deleted.  In this commit, we limit the scope of the profiler
so that its destructor is called before the queues are deleted.  Note that
it is never an issue for CUDA/HIP, because the device synchronization calls
in those backends do not need any amrex objects.  Furthermore, this is not
an issue currently for SYCL either.  But in the future, we might adopt the
WarpX approach of wrapping AMReX profilers and calling device
synchronization in the destructor.