Added a thin multipole element of arbitrary order.
A transverse momentum kick associated with a pure thin multipole.
In the future, this may be updated to allow for a sum of any
number of multipole components.

An example is included that checks the propagation of the second
moments.  In the future, this may be updated to check the preservation
of the Hamiltonian.