Examples: 1nC, no Space Charge (#212)
In ImpactX, if we set the bunch charge to zero, this means
we will have no weighting of macro particles. This might
be confusing if one wants to deposit the bunch to a density
(rho) grid.

Update all examples and disable space charge as an
effect as a whole. This is explicit, does not need a runtime
check on parallel data and will be less surprising in usage.