Fix: RFCavity Coefficients (Use-After-Free)
Fix the copy (and move) logic of `RFCavity` for the Fourier
coefficients. The challenge is that the data container that
holds the coefficients on the device can only be host-copied,
but not host-device (, device-host or device-device) copied.

Thus, we need to split our data members to keep the element
itself copyable without letting the pointers to the device
pointers go stale.