Programmable Element: Thread Safety (#325)
Expose an option to opt-in into thread safety for the programmable
element. This controls if the particle iterator loop is performed
in serial, even when the OpenMP backend is used.

Thread-safety is easy to achive for regular beamline elements that
are implemented via maps or intergrators.

Thread-safety is not easily guaranteed for interaction with external
libraries, e.g., I/O or ML packages.