Kicker Update API: Enums & Strings
Update the C++ API to use an enum for the unit system and
the Python API & inputs file syntax to accept a string. That
makes the parameters at the call sites self-describing.

Also clarifies in doc strings what dimensionless is referring
to (units of magnetic rigidity of the reference particle).