pytest: Temporary Working Directory (#424)
* Fix: pytest Wait 1 Sec

Simetimes our tests are so quick that they
finish below 1second. That does not work with
`UniqueStrings` in AMReX' `UtilCreateCleanDirectory`,
which can only do one name per second:
https://github.com/AMReX-Codes/amrex/blob/23.08/Src/Base/AMReX_Utility.cpp#L186-L199

If the target exist, the behavior of `std::rename` is
implementation dependent. We see crashes of this
on macOS.
https://en.cppreference.com/w/cpp/io/c/rename

* Use Temporary Working Directory

Avoid clean-up needs by using a cwd that is unique per test.

* All the yielded results in the cwd context

* Set CWD of runner