MAD-X Reader: Avoid Scipy
Currently, we do not need `scipy` outside of examples. This is
neat, because it is a sometimes hard to install dependency.

This updates `madx_to_impactx.py` to avoid requiring scipy for
a few constants.

In the future, if we have a hard dependency on `scipy`, we should
add it in `requirements.txt` of the project root.