Dist None: Set to Zero (#465)
With GCC 12, compiling the `None` distribution throws warnings in
compilation of `add_particles`. Originally, this was added so that
the `std::variant` for `impactx::distribution::KnownDistributions`
creates a default constructor for the variant.

Adding an assert that the distribution cannot be used in
`add_particles` does not silence the warning diagnostics.
Thus, we simply make it a distribution that sets the values to
zero.