Add 6D minimum and maximum values to reduced beam characteristics (#474)
* Add Min and Max values for 6D coords to output

Added 6D minimum and maximum values for the beam to output
of reduced beam characteristics.

* Add documentation for quantities and units

Added documentation for the min/max values to the documentation and
added unit information where it was missing for the
reduced_beam_characteristics function.

* Merge reduction operation with existing one

The reduction over all particles is now merged with the existing one
for the mean values. This avoids that we have to iterate over all
particles an additional time.