Programmable Element: Profile `pc` Push
If he `m_push` member is overwritten by the user, then
we do not call into `push_all` and thus do not add
element-wise profiling to the `Programmable` element yet.

This adds profiling for this case, which is then reflected
in the tiny profiler as well as NVTX range marks.