Tests: Mark `thermal` as slow
Mark the `thermal` test as slow.
This test runs 50sec in release mode, >10min in Debug mode.

Exclude in CI Debug builds.
Exclude in UBSAN/ASAN tests.

The thermal distribution implementation should be sufficiently covered
by the faster running `bithermal` tests and continues to be run
in the other tests in CI.