Add initialization from Twiss parameters
This changes the initialization of the beam particle distribution to allow for
Twiss parameters to be supplied.
The code first checks if the name for the supplied distribution has a suffix
indicated by an underscore. It then checks if the base name of the distribution
is an element of the list of allowed distributions that share the same input
logic. Then it chooses to initialize the beam ellipse parameters either
directly from the given inputs, or it calculates them from the Twiss parameters
and process with initialization.