Add initialization from Twiss parameters (#527)
* Add initialization from Twiss parameters

This changes the initialization of the beam particle distribution to allow for
Twiss parameters to be supplied.
The code first checks if the name for the supplied distribution has a suffix
indicated by an underscore. It then checks if the base name of the distribution
is an element of the list of allowed distributions that share the same input
logic. Then it chooses to initialize the beam ellipse parameters either
directly from the given inputs, or it calculates them from the Twiss parameters
and process with initialization.

* Add documentation and tests

Added two tests, one for distribution generation and one for the simple
FODO cell that use the new input from Twiss parameters.
Added documentation on our units and the two beam input systems.
Added a phase space ellipse figure.

* Implement reviewer comments

- change logic from "has underscore" to "ends in `_from_twiss` or `_from_cs`"
  for distribution name
- replace `amrex::Abort` with `throw std::runtime_error` to catch it in Python
- changed naming of `setDistributionParameters...`
  to `set_distribution_parameters...` to make it more similar to Python methods

* Implement more reviewer comments

- remove some newlines
- add `std::` to math functions where I overlooked it