Amend Notices for Non-Endorsement
Appended Text:

Reference herein to any specific commercial product, process, or service
by trade name, trademark, manufacturer, or otherwise does not necessarily
constitute or imply its endorsement, recommendation, or favoring by the
United States Government, The Regents of the University of California, or
Lawrence Berkeley National Laboratory.

The views and opinions of authors expressed herein do not necessarily
state or reflect those of the United States Government,
The Regents of the University of California, or
Lawrence Berkeley National Laboratory, and shall not be used for advertising
or product endorsement purposes.